backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
groq
python-multipart
python-dotenv